requirements.txt
streamlit
ccxt
pandas
pandas_ta
yfinance
python-telegram-bot
requests